+++ app/src/main/java/com/example/energii/koszt/ui/rooms/GenerateCharts.java
@@ -69,7 +69,7 @@ public class GenerateCharts {
 
         if (cursor.getCount() > 1) {
             while(cursor.moveToNext()) {
-                pieEntry.add(new PieEntry(cursor.getInt(1), String.format("%."+ numberAfterDot +"f",((float)cursor.getInt(1) / 1000)) + " kWh"));
+                pieEntry.add(new PieEntry(cursor.getLong(1), String.format("%."+ numberAfterDot +"f",((double)cursor.getDouble(1) / 1000)) + " kWh"));
                 barEntries.add(new BarEntry(labelNumberIndex, Float.parseFloat(String.format("%."+ numberAfterDot +"f", cursor.getFloat(2)).replace(",","."))));
                 colorList.add(cursor.getInt(3));
                 deviceName.add(cursor.getString(0));
@@ -83,7 +83,7 @@ public class GenerateCharts {
 
             colorList.add(cursor.getInt(3));
 
-            barEntries.add(new BarEntry(labelNumberIndex, Float.parseFloat(String.format("%."+ numberAfterDot +"f", cursor.getFloat(2)).replace(",","."))));
+            barEntries.add(new BarEntry(labelNumberIndex,Float.parseFloat(String.format("%."+ numberAfterDot +"f", cursor.getDouble(2)).replace(",","."))));
 
             deviceName.add(cursor.getString(0));
         }
@@ -228,8 +228,10 @@ public class GenerateCharts {
         if (cursor.getCount() > 1) {
             while(cursor.moveToNext()) {
 
-                pieEntry.add(new PieEntry(cursor.getInt(1), cursor.getString(0).replace("_"," ") + " " + String.format("%."+ numberAfterDot +"f",((float)cursor.getInt(1) / 1000))));
-                barEntries.add(new BarEntry(labelNumberIndex, Float.parseFloat(String.format("%."+ numberAfterDot +"f",cursor.getFloat(2)).replace(",","."))));
+                pieEntry.add(new PieEntry(cursor.getLong(1), cursor.getString(0).replace("_"," ") + " " +
+                        String.format("%."+ numberAfterDot +"f",(cursor.getDouble(1) / 1000))));
+
+                barEntries.add(new BarEntry(labelNumberIndex, Float.parseFloat(String.format("%."+ numberAfterDot +"f",cursor.getDouble(2)).replace(",","."))));
                 roomName.add(cursor.getString(0).replace("_"," ") + " ");
                 colorList.add(cursor.getInt(3));
                 labelNumberIndex++;
@@ -238,8 +240,10 @@ public class GenerateCharts {
         }else if(cursor.getCount() == 1){
 
             cursor.moveToFirst();
-            pieEntry.add(new PieEntry(cursor.getInt(1), cursor.getString(0).replace("_"," ") + " " + String.format("%."+ numberAfterDot +"f",((float)cursor.getInt(1) / 1000)) ));
-            barEntries.add(new BarEntry(labelNumberIndex, Float.parseFloat(String.format("%."+ numberAfterDot +"f",cursor.getFloat(2)).replace(",","."))));
+            pieEntry.add(new PieEntry(cursor.getInt(1), cursor.getString(0).replace("_"," ") + " " +
+                    String.format("%."+ numberAfterDot +"f",((float)cursor.getInt(1) / 1000)) ));
+            barEntries.add(new BarEntry(labelNumberIndex, Float.parseFloat(String.format("%."+ numberAfterDot +"f",
+                    cursor.getDouble(2)).replace(",","."))));
             roomName.add(cursor.getString(0).replace("_"," ") + " ");
             colorList.add(cursor.getInt(3));
